Allow multiple processes to transform a same image
The idea is to divide all processes in group of the same
size where each group is responsible for the processing
of the same images. Each group has a master, responsible
for gathering its group's images to all the slaves in its
group. In the end, slaves of the same group send their work
to their masters, who will send the complete transformed
message to the Dungeon Master (rank 0 in MPI_COMM_WORLD).

In the current implementation, only the Sobel filter is
parallelized using the slaves. All other filters are applied
only by the masters because we didn't find any better
strategy so far.